Renamed bibleIndex to languages
